StackCollections.USet — Type

USet{U <: Unsigned} <: AbstractSet{UInt32}

An immutable, sorted bit set backed by an integer of type U. Can contain the integers UInt32(0):UInt32(B - 1) when backed by an integer consisting of B bits.

Construct from an iterable of integers, or an empty set with USet{U}().

Operations that would introduce an unrepresentable member, or require a nonempty set when given an empty one, throw an ArgumentError. These checks can sometimes be disabled locally with @inbounds.

intersect and setdiff with a generic collection accept Real and Complex elements, and ignore those that are not isequal to a representable member, such as 0.5, -0.0 or 1.0 - 0.0im. Other element types throw a MethodError.

Mutating operations are not supported; use push, pop and popfirst instead of the corresponding mutating Base operations.

The memory of this type is guaranteed to be identical to a U, where set bits from LSB to MSB represent the presence of the integers zero and upwards. E.g. USet{UInt8}([0, 3, 5]) is guaranteed to have the same memory layout as 0x29. Obtain the equivalent integer with to_bits. This is guaranteed to be a noop. Construct from the equivalent integer with from_bits.

Examples

julia> s = USet{UInt32}([4, 2, 1]);

julia> collect(s) == [1, 2, 4] # ordered
true

julia> ss = push(s, 5); ss == s # immutable
false

julia> symdiff(s, ss) == USet{UInt8}(5)
true

StackCollections.UVec — Type

UVec{U <: Unsigned} <: AbstractVector{Bool}

Immutable boolean vector backed by a single U. A T <: UVec has a maximum length determined by U, which can be queried by capacity(T).

Construct from an iterable of elements convertable to Bool.

Extended help

Operations that exceed the maximum capacity, or require a nonempty vector when given an empty one, throw an ArgumentError. Invalid indices throw a BoundsError. These checks may sometimes be disabled locally with @inbounds.

Mutating operations are not supported; use push, pushfirst, pop, popfirst, deleteat, and setindex instead of the corresponding mutating Base operations.

The underlying integer of an x::UVec can be obtained with to_bits, and a UVec can be constructed from its underlying integer with from_bits.

Most operations on UVec that return boolean vectors, such as filter, reverse and indexing, are specialized to return UVec. However, it is not guaranteed that all methods are implemented and do not fall back to a default implementation. However, implemented specialized methods that explicitly return UVec are guaranteed to not be removed in future minor releases. Base.similar(::UVec, args...) returns BitArray.

StackCollections.append — Method

append(v::UVec{U}, itr)::UVec{U}

Convert each element of itr to Bool, and push them, in order, to a new copy of v, which is returned. Throw an ArgumentError if the appended elements can't fit in a UVec{U}; this may be suppressed with @inbounds.

See also: push, insert

Examples

julia> v = UVec{UInt16}([1, 0]);

julia> v2 = append(v, (i for i in [1, 0, 0, 1]));

julia> v2 == v
false

julia> v2 === UVec{UInt16}([1, 0, 1, 0, 0, 1])
true

StackCollections.can_contain — Method

can_contain(::Type{<:USet{U}}, i::Union{Real, Complex})::Bool

Return whether a USet{U} can contain an element isequal to i, by checking if i is an integer in 0:maximum_member(USet{U}). Values that are numerically equal to such an integer, but not isequal to it, such as -0.0 or 1.0 - 0.0im, cannot be contained.

julia> can_contain(USet{UInt32}, 55)
false

julia> can_contain(USet{UInt64}, 55)
true

julia> can_contain(USet{UInt8}, 2.0 + 0im)
true

julia> can_contain(USet{UInt8}, -0.0)
false

StackCollections.capacity — Method

capacity(::Type{<:UVec{U}})::Int

Compute the maximum number of elements a UVec{U} can contain. This computation is a compile-time constant.

Examples

julia> capacity(UVec{UInt8})
5

julia> capacity(UVec{UInt32})
27

StackCollections.deleteat — Method

deleteat(v::UVec{U}, idx::Integer)::UVec{U}
deleteat(v::UVec{U}, idx::UnitRange{<:Integer})::UVec{U}
deleteat(v::UVec{U}, mask::AbstractVector{Bool})::UVec{U}

Return a new UVec based on v, but with the index or indices idx removed, and all subsequent elements shifted downwards to fill the deleted elements.

If a Boolean vector mask is passed, including a UnitRange{Bool}, it is used as a mask, and the elements at positions where mask is true are removed.

Throw a BoundsError if idx is out of bounds for v, or if mask does not have the same length as v. This can be disabled with @inbounds. Throw an ArgumentError if idx is a Bool, even with @inbounds.

See also: pop, popfirst, spliceinto

Examples

julia> v = UVec{UInt8}([1, 1, 0, 1]);

julia> deleteat(v, 2:3) |> print
Bool[1, 1]

julia> deleteat(v, 4) |> print
Bool[1, 1, 0]

julia> deleteat(v, [true, false, false, true]) |> print
Bool[1, 0]

julia> deleteat(v, 4:5) |> print
ERROR: BoundsError: attempt to access 4-element UVec{UInt8} at index [4:5]
[...]

StackCollections.from_bits — Method

from_bits(::Type{USet}, u::U)::USet{U} where {U <: Unsigned}

Construct a USet{U} using the backing integer u. For bitstype U, it is guaranteed that to_bits(from_bits(USet, u)) === u.

The resulting USet contains the elements represented by the set bits in u, with UInt32(0) being the LSB, and UInt32(bitsizeof(U) - 1) being the MSB.

This operation is guaranteed to be optimizable to a noop.

julia> u = 0x50ae; s = from_bits(USet, u);

julia> s isa USet{UInt16}
true

julia> s == Set([i for i in 0:16 if isodd(u >> i)])
true

julia> to_bits(s) === u
true

StackCollections.from_bits — Method

from_bits(::Type{UVec}, u::U)::UVec{U} where {U <: Unsigned}

Construct a UVec{U} from its underlying integer. Not all integers are valid backing storage for a UVec, so the result may be a corrupted and malfunctioning UVec. The only way to get a guaranteed valid input to this function is to_bits(::UVec).

This function is guaranteed to be:

Optimizable to a noop

Round-trippable with to_bits, in the sense that for any v::UVec, from_bits(UVec, to_bits(v)) === v.

Examples

julia> v = UVec{UInt32}([0, 1, 1, 1, 0, 1, 0, 1, 0, 1]);

julia> u = to_bits(v); typeof(u)
UInt32

julia> from_bits(UVec, u) === v
true

StackCollections.insert — Method

insert(v::UVec{U}, idx::Integer, item)::UVec{U}

Convert item to Bool, then return a new UVec{U} based on v, but with the converted item inserted at index idx. The elements at or after idx are shifted one index up. The index idx must be in 1:length(v)+1. Throw a BoundsError if idx is out of bounds. Throw an ArgumentError if v is at capacity. Both are disabled with @inbounds. Throw an ArgumentError if idx is a Bool, even with @inbounds.

See also: spliceinto, push, deleteat, append

Examples

julia> v = UVec{UInt8}([1, 1, 0, 1]);

julia> insert(v, 2, 0) |> print
Bool[1, 0, 1, 0, 1]

julia> insert(v, 5, 1) |> print
Bool[1, 1, 0, 1, 1]

julia> insert(v, 6, 1) |> print
ERROR: BoundsError: attempt to access 4-element UVec{UInt8} at index [6]
[...]

StackCollections.pop — Function

pop(collection::Union{USet, UVec}) -> (new_collection, item)

Create a new collection of the same type as the input, but with the last element removed. The new collection and the removed element are returned as a tuple.

If the collection is empty, throw an ArgumentError. The check can be disabled locally with @inbounds, similar to BoundsErrors.

See also: popfirst, push, deleteat

Examples

julia> (v2, element) = pop(UVec{UInt16}([1, 0]));

julia> element
false

julia> v2 === UVec{UInt16}([1])
true

StackCollections.pop — Method

pop(s::USet{U}, i::Integer)::Tuple{USet{U}, UInt32}

Return (new_s, item), where item is equal to i, and new_s is a copy of s but with item removed.

Throw an ArgumentError if i is not in s. This error may be suppressed with @inbounds, in which case this operation may give nonsensical results.

julia> s = USet{UInt8}([0, 2, 3, 6]);

julia> (new_s, item) = pop(s, 2); item === UInt32(2)
true

julia> new_s === typeof(s)([0, 6, 3])
true

julia> pop(s, 99999)
ERROR: ArgumentError: Item 99999 not present in USet{UInt8}
[...]

StackCollections.popfirst — Function

popfirst(collection::Union{USet, UVec}) -> (new_collection, item)

Create a new collection of the same type as the input, but with the first element removed. The new collection and the removed element are returned as a tuple.

If the collection is empty, throw an ArgumentError. The check can be disabled locally with @inbounds, similar to BoundsErrors.

See also: pop, push, deleteat

Examples

julia> (v2, element) = popfirst(UVec{UInt16}([1, 0]));

julia> element
true

julia> v2 === UVec{UInt16}([0])
true

StackCollections.push — Method

push(s::USet{U}, i1, is...)::USet{U}

Convert every element of (i1, is...) to UInt32, then return a new USet{U} based on s, but with the converted elements added.

Throw an ArgumentError if any converted element is larger than maximum_member(USet{U}). The check can be disabled locally with @inbounds, similar to BoundsErrors. For Integer elements, this also disables the check that they can be converted to UInt32.

See also: pop, can_contain

Examples

julia> s = USet{UInt8}([1, 4]);

julia> push(s, 0x02) == USet{UInt8}([1, 2, 4])
true

julia> push(s, 0x02, 3.0, true) == USet{UInt8}([1, 2, 3, 4])
true

julia> push(s, 8)
ERROR: ArgumentError: Value out of range for USet{UInt8}: Supports 0:7, got 8
[...]

StackCollections.push — Method

push(v::UVec{U}, i1, is...)::UVec{U}

Convert every element of (i1, is...) to Bool, then return a new UVec{U} based on v, but with the converted elements, in order, appended to the end.

Throw an ArgumentError if v is already at maximum capacity. The check can be disabled locally with @inbounds, similar to BoundsErrors.

See also: pop, pushfirst

Examples

julia> v = UVec{UInt32}([1, 1, 0, 1, 0]);

julia> push(v, 0x01) == [1, 1, 0, 1, 0, 1]
true

julia> push(v, 0x01, 0, true) == [1, 1, 0, 1, 0, 1, 0, 1]
true

StackCollections.pushfirst — Method

pushfirst(v::UVec{U}, i1, is...)::UVec{U}

Convert every element of (i1, is...) to Bool, then return a new UVec{U} with the content of v, but with the converted elements, in order, at the beginning, and all preexisting elements shifted back. Throw an ArgumentError if v is too big to accommodate the extra elements (see capacity). The check can be disabled locally with @inbounds, similar to BoundsErrors.

See also: push

Examples

julia> v = UVec{UInt8}([1, 0, 1, 1]);

julia> v2 = pushfirst(v, true); v2 == [1, 1, 0, 1, 1]
true

julia> v == v2
false

julia> pushfirst(UVec{UInt16}(v), true, true, false) == [1, 1, 0, 1, 0, 1, 1]
true

julia> pushfirst(v2, false)
ERROR: ArgumentError: UVec at maximum size
[...]

StackCollections.setindex — Method

setindex(v::UVec{U}, item, indices::Integer)::UVec{U}
setindex(v::UVec{U}, items::AbstractVector, indices::AbstractVector{<:Integer})::UVec{U}

Return a new UVec{U} based on v, but with the elements at indices set to items. Each element of items is converted to Bool and assigned to the corresponding index in indices. For repeated indices, the last assignment wins.

Boolean vector indices are masks: their length must match v, and items must contain count(indices) items. Scalar Bool indices throw an ArgumentError, even with @inbounds.

Throw a BoundsError if any index is not an existing index of v. Else, if the number of selected indices does not match the length of items, a DimensionMismatch error is thrown. These errors may be elided with @inbounds.

See also: push, spliceinto

Examples

julia> v = UVec{UInt64}([1, 1, 1, 1, 1, 1, 1, 1, 1]);

julia> v2 = setindex(v, [0, 1, 0, 1, 0], 3:7); println(v2)
Bool[1, 1, 0, 1, 0, 1, 0, 1, 1]

julia> v2 == v # v is unchanged
false

julia> setindex(v, [1, 0, 1], 2:3)
ERROR: DimensionMismatch: Tried to assign 3 items to 2 indices
[...]

StackCollections.spliceinto — Method

spliceinto(v::UVec{U}, i::Integer, e::AbstractVector)::UVec{U}

Return a copy of v with the elements of e (converted to Bool) inserted at positions i:i+length(e)-1, and the previously existing elements at i:length(v) shifted to higher indices.

Throw a BoundsError if i is not in 1:length(v)+1. Throw an ArgumentError if the resulting UVec{U}'s length would exceed capacity(UVec{U}). Both exceptions can be suppressed with @inbounds.

Boolean indices are not supported: Throw an ArgumentError if i is a Bool, even with @inbounds.

See also: insert, deleteat, push

Examples

julia> v = UVec{UInt32}([0, 1, 1, 1, 1, 0, 0]);

julia> v2 = spliceinto(v, 3, [0, 0, 0]);

julia> v2 === typeof(v)([0, 1, 0, 0, 0, 1, 1, 1, 0, 0])
true

julia> spliceinto(v, 8, [1, 1]) == [0, 1, 1, 1, 1, 0, 0, 1, 1]
true

StackCollections.spliceinto — Method

spliceinto(v::UVec{U}, i::UnitRange, e::AbstractVector)::UVec{U}

Return a copy of v with the indices at i deleted, and the elements of e (converted to Bool) inserted at the deletion site. Elements in v after i are shifted to immediately after the inserted e.

Throw a BoundsError if i is not in 1:length(v). Throw an ArgumentError if the resulting UVec{U}'s length would exceed capacity(UVec{U}). Both exceptions can be suppressed with @inbounds.

Boolean indices are not supported: Throw an ArgumentError if i is a UnitRange{Bool}, even with @inbounds.

This is equivalent to spliceinto(deleteat(v, i), first(i), e), but more efficient.

See also: insert, deleteat, push

Examples

julia> v = UVec{UInt32}([0, 1, 1, 1, 1, 0, 0]);

julia> v2 = spliceinto(v, 3:4, [0, 0, 0]);

julia> v2 === typeof(v)([0, 1, 0, 0, 0, 1, 0, 0])
true

julia> spliceinto(v, 2:6, [1, 1]) == [0, 1, 1, 0]
true

StackCollections.to_bits — Method

to_bits(x::USet{U})::U

Get the integer backing x. This integer is guaranteed to be a bit mask where the N-th zero-indexed bit from lowest to highest bit represents the presence of the element N in x.

See also: from_bits

Examples

julia> s = USet{UInt16}([4, 11, 8, 2, 6]);

julia> u = to_bits(s);

julia> u === 0x0954
true

julia> from_bits(USet, u) === s
true

StackCollections.to_bits — Method

to_bits(v::UVec{U})::U

Obtain the backing integer of v. The value of to_bits(v) is an implementation detail, but the following properties are guaranteed:

This operation can optimize to a noop

The result is a value of type U

No distinct UVecs map to the same integer; however, the inverse is not guaranteed.

For any v::UVec, from_bits(UVec, to_bits(v)) === v.

See also: from_bits